moved handlers out of server.py
